Work on datadriven omx.
Make it possible to run the openmx datadriven test specifying openmx
keywords at a lower level (scf_criterion=x instead of the corresponding
python-name).  Also, we will need to accept write() functions that take
a parameters={..} argument as well as those which take **kwargs.
I think we should accept only the former option in the future.