ENH: added table printing for calculator outputs
Borrowing the (undocumented) ase.gui.images Images object to print out
results similarly to how the gui displays a plot. The table becomes
very wide if both displacements and forces are in the same one and so
different tables are printed. But both of these quantities should generally
be compared together. A briefer table of only magnitude differences
in positions and forces may be added.