ENH: added rank ordering and show only n lines
The atoms in each pair difference of iterations can now be rank ordered
and the number of atoms to display can be changed. The cli only shows
differences between two atoms or images, but for those cases where the
magnitude at each step is also relevant, it displays the average (in the case of force)
or both values (in the case of energy).

 There is a possibility to use verbosity to control the table
width. It may be desirable to select which columns to display by the
user rather than having a counter verbosity option. This requires changing
which number of rows containing the data to pass to the numpy
transpose which makes the table and making the row formatters change in
the corresponding way.