ENH: general table formatting and sorting
Changed to support a general table creation including left-to-right
order of columns and top-to-down order of rows by (hierarchial) sorting
on field(s). Added support for two trajectory files to compare the same
ith image of the two. Preliminary support of user customization.
Non-numeric fields (element symbols) are dealt with by mappings and custom string
formatter class and method.

To do:
. tests of all use cases
. add verbosity options to default fields
. support rank order option again
. support show lines option again
. improve string formatting dictionaries (now nested dictionaries) in what is now called template.py