DOC: _format_action is overridden but invocations
The simple indent problem defining it not as a class method but a
function within the class method _fill_text was fixed. Also fixed the
whitespace correction of the input to strip only after the new lines
from textwrap to not have words sticktogether. Though the string
returned by _format_action is desirably formatted, there is a chain of calls
including invocation functions which change the result.