Refactor Vibrations tests into unittest
This will probably be replaced in the near future once pytest is
figured out, but for this kind of refactoring work it's really helpful
to have a proper test framework with a bit more structure in the test
files. Use the same kind of hackery as XDATCAR tests to integrate
unittest into the ASE test framework.