VibrationsData: indexing/frozen atoms, more tests
VibrationsData should now work with data including frozen
atoms. Another test example has been added; N2 adsorbed to a frozen Ag
surface. The BFGS optimisation of this structure takes a few seconds
so the optimised structure has been dumped into the file, and the
appropriate code is included as a comment to assist future changes.

The appropriate way to handle frozen atoms is to simply omit them from
the Hessian; they are not part of the vibrating system. While they are
accepted as indices for consistency, it seems nicer to use a boolean
mask internally.