TST: Test Vibrations summary output. Reduce stdout pollution
The vib.py test writes a lot of stuff to STDOUT. ASE's test framework
hides some of that, but it's still annoying when e.g. coverage
testing. Here most of the output is suppressed by writing to files
that can be checked (if appropriate) and then deleted.

We still get the list of .pckl files written out, but that won't last
forever and I don't want to add any logic to that object while
in the process of replicating/replacing it.