STY: Flake8 vibrations, including whitespace in vib.summary()
Testing vib.summary() was a bit awkward as it needs to be compared
with a text block that includes trailing whitespace. While it's
possible to supress the flake8 errors and remember to never

    M-x whitespace-cleanup

(or whatever vim users do), it seems cleaner to just remove
the (unnecessary) trailing whitespace from the summary() output.