VibrationsData: caching, JMOL XYZ output with ase.io extxyz writer
As the underlying data are now immutable, we can safely cache the
solution of the eigenvalue problem - it cannot become stale.  This
cleans up some functions as the option to pass precomputed results is
eliminated and caching becomes a hidden implementation detail.

With this caching, separate get_modes() and get_energies() functions
become convenient without wasteful recalculation.

The implementation of eigenvalue problem is slightly modified so that
the multiplications by mass weights are done in _exactly_ the same way
as the Vibrations object. Numerical differences in the 20th decimal
place were frustrating testing/validation between the methods.

Move mass-weighting into get_energies_and_modes(); "modes" should
always be mass-weighted. If we want a pure eigenvector, call it
"eigenvector" to make that clear, or "vector" if
lazy/horizontally-constrained.

Use the existing extxyz writer to cut down redundancy and make
better/more extensive use of ase.io, in the spirit of current
ASE development priorities.
This does raise one tricky issue: by using the extxyz writer we
lose control over what is written in the image line (unless it's going
to be the same comment for every line). This is good because it means
ASE stands a decent chance of reading the data back in again to a
series of images with metadata. But it also means that the information
given in the JMOL user interface is less human-friendly and may even
exceed the visible region.