Vibrations flake8 tidy-up
There are two remaining warnings regarding use of capitals in variable
names. In both cases these correspond to names exposed in the API (a
.H attribute for a Hessian and a kT argument for thermodynamic
temperature).

Leaving these unchanged in the existing object to avoid compatibility
breakage. Flake8-compliant names are/will be used in new
implementation so when Vibrations class is removed, so is the
non-compliance.