Tighten up typing in ase.vibrations.vibrations using mypy
mypy identified a few unpleasant typing scenarios, so tweak
VibrationsData: indices are always stored in self._indices as a
list (or None) and self.indices @property always returns a
np.array (or None).

Sanity checks added to fromdict() using bare assert statements for
visibility to mypy. It is fundamentally tricky to use heterogeneous
dictionaries with mypy at the moment as there is not an appropriate
syntax in Python standard library to express typing of specific
dictionary keys. Advice on stackexchange etc. is "use a NamedTuple"
or, worse, "just make another class". This is clearly not viable for a
fromdict() method, and in any case would be a severe compromise in
code simplicity. Hopefully an appropriate syntax can be adopted soon?