DEV: refactor of table template filling to object
The table template filling code is simplified by making a table object which
takes the options which don't change on iterating through the images in
its initialization and has a method to make the string for the table
when passed a pair of atoms objects. Only one table with a list of
summary functions is made depending on if the calculations exist or not.
More default summary functions and command line customizable summary
functions may be useful to develop.

The same slice syntax as in ase gui cli is supported.