DEV: OOP refactor of table making
The majority of the program is now object oriented in the Table object.
There is more control of formatting and better default formatting in the
initialization of the class which has both more inputs and conditionals on None
type inputs.

The get data functions were made more brief and descriptive by defining
deltas for the fields that use them.