Add get_spreads() for single function spread
Add a new method to get information about the spread of every Wannier
function in the set. The previous method get_radii() returns a null
radius with some lattice types (a.e. bulk Si and GaAs). This new method
should be more consistent but it is defined as a relative spread for
functions inside a set, the value cannot be used to compare two
different sets of Wannier functions.