DEV: added fields for non-difference images data
The fields are position and forces for both images. It is possible to
use the diff command as a display command by substituting a hypen-minus
for one file name (and it is easy enough to make an alias for this if it is expected to be repated).

Other than than addition to the list of possible
fields there is a change of case structure for brevity and fewer
redundant evaluations, documentation.