STY: pyflakes and pylint
Removed unneeded module imports, edited white space and line length to
pep8 standard, changed module import call to template in diff.py for
the used objects directly, changed Table header like body to no longer
be an attribute but a local variable in the make method.

Large number of branches and the main Table class with a large number of
initialization arguments remain.