DEV: combine get data functions
The separate functions have been combined. The list of atoms properties,
and the branch of atoms properties or calculator properties, is kept.
This is because it is best to have as many classifications of the properties
as possible to make a decision tree with the first two nodes being atoms data
 or calculator data. However, the first check does an exact match and
then redundant matches and partial matches. Given the conventions, e.g.,
to prepend 'a' for average, a decision tree based on position in string
isn't possible.