Add automatic guess of number of Wannier functions
A new function can help the user at guessing the right number of Wannier
functions for a system, trying different values and choosing the one
for which the spread of the most delocalized WF is minimum.

Fix the generation of a random unitary matrix, it does not make sense to
generate a rotation matrix and change the data type in the case of
complex valued matrix.