ENH: Expand testing with two-body forces and stress for ase.calculators.lj
Now covers cases with non-default parameters and at semi-random positions,
as well as physics-based sanity checks to aid in troubleshooting errors.

This will not yet catch any errors due to partitioning, since for two atoms
all the combinations collapse to one. The next step is therefore to expand the
tests to three atoms as well. It also doesn't yet cover periodic cases.