BUG: need to raise error for unknown keywords
This calculator only incorporates a limited set of keywords
into the input file that is sent to CP2K, and any other
keywords do not have any effect. If a user employs a keyword
that is not in this list (e.g. 'kpts=...'), then that should
raise an error. Otherwise the user might think that the keyword
is legitimate and would have a certain effect.