dos.sample() to dos._sample() -> array; fix dos.sample_grid() scale
Sampling at arbitrary energies is useful as an underlying method, but
there is no guarantee that the resulting data will constitute a
sensible DOS. To emphasise this:

- change the method name to _sample() in both DOSData and DOSCollection
- change return type to a simple array of weights

When resampling data that has already been broadened, the bin-width
should be accounted for (treating the input data as a
histogram). Otherwise, a DOS with fine x-axis sampling with gain a
higher intensity as each a unit Gaussian is added to initial point.

- multiply the output weights of sample_grid() by the initial
  bin-width when resampling from a GridDOSData.

Funnily enough this passes the existing plot test, because the test
data has a bin width of 1. A more rigorous resampling test should
therefore be added; that is not included in this commit.