BUG: fix Fermi level parsing for DFTB+ v20.1
Until now, the Fermi level parser always expected 2 Fermi levels,
as this was how DFTB+ v17.1 behaved (with or without spin polarization).
The Fermi level(s) were then removed from the results.tag file in v18.1
(as were the eigenvalues). Upon their re-introduction in v20.1, only one
Fermi level is provided in the absence of spin polarization.
With this fix the DFTB+ tests no longer fail with the 20.1 version.