MAINT: add_isosurface_to_pov updated, made method
string formatting updated to python 3.6 for add_isosurface_to_pov.
tested that it reads in the data and the image is created for an
isosurface.

not yet made pep compliant

write_pov is made backwards compatible through the general api
ase.io.write pending some consensus on the desirable api for povray