added optional addition of "region" array to atoms
Region array is used to store mapping of the cell between QM, buffer and MM regions. This can be used for visualisation of the regions in the cell as well as for validating the mapping. If the original atoms object already has a "region" array, the new regions will be added to the existing array.