ENH: VibrationsData with_new_mass() for isotope studies
Many atomistic methods do not distinguish between isotopes when
evaluating forces/energies. Within this approximation we can explore
the impact of isotopic substitutions on the vibrational spectrum by
modifying atomic masses without recalculating the Hessian.

It would be convenient to do this by modifying a VibrationsData
in-place. As VibrationsData deliberately lacks the mutability required
for this, we provide a convenience function to support this use-case
by spawning a new VibrationsData object.