Vibrations: add "no-op" np.asarray so hinted types will work
Sequence[complex] does not have a .real method so needs to be cast to
array for this hint to be correct. We provide this hint so that it is
easier to reason about what is being passed around; np.ndarray can be
almost anything!

In real-world usage the argument will always be an np.ndarray and so
np.asarray() is cheap.