refactor to composite plottingVariables and POVRAY
povray takes as input some of the attributes from plotting variables.
it is somewhat like a composite object, but it assigns all attributes as
its own. therefore better might be to have the POVRAY object take as
parameters inputs, rather than an object which it takes the attributes
from

current commit is broken for gui