gui is fixed
separate plottingvariables and povray dictionaries are passed. reverted
back to a single write_pov function call. this is for testing purposes
for all examples which expect a single write_pov call.

since the povray class, on instantiation, offsets the positions and
reshapes the cell vertex, awkwardly the isosurface data has to be
initialized afterward with the attributes of the povray class to which
it will be assigned as an attribute. this can be avoided
or hidden from the user by having an instance method on povray which
makes the isosurface instance