added class methods for construction by objects
this allows the initialization to be independent of other objects, so
other object changes only require the relevant constructor be changed.
also makes things convenient for the user since they don't have to be
aware of the data passed from objects

one observation: now that some calculations have been moved to the
initialization of the POVRAYIsosurface object, one use case, of resetting the cut-off
value of the object, can no longer be used. it would be possible to
define a setter for the cut-off value that recalculates the required
quantities