removed qm_cluster_cell, qm_cluster_pbc and qm_shift from properties
The cell and pbc is stored in qm_cluster atoms object. Thus the function get_qm_cluster() only returns the cluster and does not set anything. This is more consistent with the function name. The function is called and the qm_cluster is initialised on every call of calculate() function. This is the desired behaviour for the simulations when the qm region and thus qm cluster may change during simulation.