ENH: Add smooth cutoff option to ase.calculators.lj
Adds a jax-md / HOOMD-blue style smooth cutoff to the LJ calculator.

Added a bit more documentation to the docstring so when someone touches
the calculator next time, they can hopefully figure out the somewhat tricky
changes to forces etc. that need to be done to get the smooth cutoff to work.

Also added testing for this case.

I also externally tested this against the jax-md implementation, and against
finite differences to check that the pairwise forces are actually the correct
derivative. The reference pressure test fails with the smooth cutoff, but not
by much -- I think this is ok, given that the smooth cutoff does indeed modify
the PES somewhat.