ENH: Read more sections of a Gaussian input file
Previously, only the atom symbols and positions could be read from
an input file. Added in GaussianConfiguration to allow reading in
of the route section, link0 section, charge and multiplicity.
The read_gaussian_in method still only reads in the atoms from the
input file and returns an atoms object, but you can get the other
parameters through the GaussianConfiguration object.