ENH: Update write_gaussian_in
Updated write_gaussian_in so that it is compatible with
the case when we have read in the parameters from the
input file.
Uses the output type saved to the parameter: output_type
instead of setting it to 'P' always.
Does not set a default method - this is because when
reading the input file, the method may not have been
saved to the parameters dictionary under 'method',
but it will still be written to the new input file.