ENH: Reads and writes nuclear properties
Reads in from a gaussian input file any nuclear
properties contained in brackets after the element
symbols.
Also reads in temperature, pressure, scale and masses
in the ReadIsotopes section if present.
Saves the masses as the masses of the Atoms object.
Writes out the nuclear properties and ReadIsotopes
details when generating an input file.