MAINT: Make function for model refresh
Because cutoffs can generally be just another parameter registered by
the model, we need to ensure that any time a model's parameters are
updated, its clear_then_refresh() function gets called at some point
before its compute() function is called.  In addition, we need to ask
the model for its updated influence distance and cutoffs as soon as this
is done and use them to update our neighbor lists.