BUG: Reimplement check_state
check_state() needs to take into account whether model parameters have
changed or not, since otherwise calculate() is generally not even going
to be called.  Also, the reimplementation of calculate() needs to call
the original Calculator.calculate() so that the Atoms object will be
cached for use in compare_atoms().