Vibrations bugfix/tidy-up: Improve handling of frozen atoms
A bug was reported when using write_jmol with frozen atoms #895

For mathematical analysis of the displacement vectors, we generally
want the array to correspond to the frequencies and active atoms.

For visualisation we generally want all the atoms to be included (with
zeroed entries for the inactive atoms.)

Rather than implement this conversion in multiple places (and
occasionally get it wrong or forget to include it), we can centralise
that conversion with an intuitive "all_atoms" parameter to
VibrationsData.get_modes(). It is then simple to restore the
legacy (all-atoms) convention of Vibrations.get_mode().