ENH: Only call KIM model refresh when necessary
Mutating any one of a KIM portable model's parameters generally results
in potential changes to its other parameters.  It also generally means
that the quantities (the cutoffs, more or less) used to construct
neighbor list(s) may also change.  As a result, one must ask the model to
perform these consistency updates (and propagate any changes to the
neighbor list-related parameters to whatever code actually constructs
the neighbor list(s)).  Prior to this commit, this update process was
done every time a call to KIMModelCalculator.set_parameters was made.
However, it only really needs to be done when either (1) someone
requests the values of the model parameters or (2) the calculator's
calculate() function is called to compute energy/forces/etc.