TST: add test_plumed to ase/test to check the new calculator Plumed
This code contains 4 tests. (1) test_CVs calls plumed-ASE calculator
for computing some CVs and computes those CVs directly from
atoms.get_positions() for comparing them. (2) test_metadyn computes
a Metadynamics calculation, this result is compared with the same
calculation made externally. (3) test_restart is the same that the last
one, but restarting the simulation. (4) test_postpro takes the trajectory
created un test_metadyn for obtaining the same bias (which is history-dependent).
(5) there is an additional test called test_remove. It is there just for
deleting the files created in the previous tests.