Update espresso.py by correcting reading of magmoms for spin polarized pw.x output files.
I have observed that ASE finds it difficult to read output files generated by spin-polarized dft. It fails at this line: float(mag_line.split()[] for mag_line. it reads 'magn=' which cannot be converted to float which throws up this error. Thus, I just corrected the index to get the last element of the line (or it should be the 6th element). I have tested it and now it works properly.