MOPAC Calculator: avoid false-positive in parser
Recent versions of MOPAC require the DISP keyword to be used for the
expected TOTAL ENERGY to be written. This also writes out another
line containing 'FINAL HEAT OF FORMATION' does not contain the energy
in the expected position and causes an error.

Issue describing MOPAC change
https://github.com/openmopac/mopac/issues/66

To accommodate these changes:
- use a regex for stricter pattern matching
- always include DISP in task list